+++ tests/testthat/test-compute_ofv.R
@@ -4,7 +4,7 @@ test_that("compute basic ofv", {
     adm_lines(amt = 10, addl = 2, ii = 12) %>%
     obs_lines(DV = c(.1, .15), time = c(18, 40)) %>%
     add_covariates(list(WT = 70)) %>%
-    see_data()
+    get_data()
   arg.ofv <- c(preprocess.ofv.fix(x = mod), preprocess.ofv.id(x = mod, iddata = data1))
   of_value <- compute_ofv(eta = c(ETA1 = -.2, ETA2 = .1, ETA3 = .2),
               mrgsolve_model = arg.ofv$mrgsolve_model,