+++ tests/testthat/test-valid_dataset.R
@@ -2,7 +2,7 @@ mod1 <- mread('ex_mbr1', mbrlib())
 data1 <- mod1 %>%
   adm_lines(time = 0, amt = 100) %>%
   obs_lines(time = 20, DV = 1.5) %>%
-  see_data()
+  get_data()
 data1
 test_that("Dataset check is correct", {
   expect_error(mapbayest(mod1, data = select(data1, -ID)), "ID column is missing")
@@ -26,14 +26,14 @@ test_that("misspecification of evid/mdv are checked", {
   data2 <- mod1 %>%
     data_set(data1) %>%
     obs_lines(time = c(25,30), DV = c(1, .5)) %>%
-    see_data() %>%
+    get_data() %>%
     mutate(evid = c(1, 1, 1, 3, 4))
   expect_error(mapbayest(mod1, data = data2), "Lines with mdv = 0 must have evid = 0.")
 
   data3 <- mod1 %>%
     data_set(data1) %>%
     obs_lines(time = 0, DV = .005) %>%
-    see_data()
+    get_data()
 
   expect_error(mapbayest(mod1, data3))
 })